# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OAuth 2.0/2.1 Authorization Server API
  description: |
    Production-ready OAuth 2.0/2.1 Authorization Server with OpenID Connect support.
    
    ## Features
    - OAuth 2.0/2.1 compliant with PKCE requirement
    - OpenID Connect Core 1.0
    - Token Revocation (RFC 7009)
    - Token Introspection (RFC 7662)
    - Dynamic Client Registration (RFC 7591)
    - Device Authorization Grant (RFC 8628)
    - Multi-Factor Authentication (TOTP)
    
    ## Authentication
    Most endpoints require either a valid access token or client credentials.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://auth.example.com
    description: Production server
...
paths:
...
  /authorize:
    get:
      tags: [Authorization]
      summary: Authorization Endpoint
      description: |
        Initiates the OAuth 2.0 authorization flow. Redirects user to login
        if not authenticated, then to consent screen if consent is required.
      operationId: authorize
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be 'code' for authorization code flow
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of scopes
        - name: state
          in: query
          schema:
            type: string
          description: Random state for CSRF protection
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge (required by OAuth 2.1)
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256]
          description: Must be S256
        - name: nonce
          in: query
          schema:
            type: string
          description: Nonce for OpenID Connect
        - name: prompt
          in: query
          schema:
            type: string
            enum: [none, login, consent]
      responses:
        '302':
          description: Redirect to login, consent, or callback
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtained from the token endpoint
    clientBasic:
      type: http
      scheme: basic
      description: Client ID and secret as username:password
  schemas:
...
    OAuthError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - invalid_scope
            - unauthorized_client
            - unsupported_grant_type
            - invalid_token
            - insufficient_scope
            - access_denied
            - server_error
        error_description:
          type: string
        error_uri:
          type: string
          format: uri